requirements.txt
# requirements.txt - Versiones compatibles con Python 3.12
kafka-python==2.0.3
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.0
confluent-kafka==2.2.0
pyarrow==14.0.1